Refactored to use getFunctions runtime loader
