+++ README.md
@@ -12,6 +12,31 @@ _Twilio Flex 2.0 - VIBER Channel_ is a custom channel connector to enable Viber 
 2. Create a bot - https://partners.viber.com/account/create-bot-account
 3. Complete required fields and copy secure token
 4. Put token in `.env` file `VIBER_AUTH_TOKEN`
+5. Configure viber webhook
+
+#### Viber Webhook
+
+Once the Twilio serverless function is deployed, obtain the host name and set a POST to Viber, e.g.
+
+Note: Remember to set the `X-Viber-Auth-Token` header to the previously obtained token in step 3
+
+```json
+{
+    "url": "https://<YOUR_HOST_NAME>.twil.io/api/viber/incoming",
+    "event_types": [
+        "message",
+        "delivered",
+        "seen",
+        "failed",
+        "subscribed",
+        "unsubscribed",
+        "conversation_started"
+    ],
+    "send_name": true,
+    "send_photo": true
+}
+```
+
 
 ![Flex 2.0 - Viber Bot](docs/flex2_viber_bot.png)
