+++ src/functions/api/common/common.helper.private.ts
